requirements.txt
tqdm
xmltodict
scikit-learn
opencv-python
timm
torchinfo
tensorboard
mmengine
mmcv>=2.0.0 -f https://download.openmmlab.com/mmcv/dist/cu116/torch1.12.0/index.html
mmpretrain